# sentinel/docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SENTINEL API
  version: "1.0.0"
  description: |
    AI-Generated Code Governance & Compliance API.

    SENTINEL scans code diffs for security vulnerabilities, license compliance,
    quality issues, policy violations, dependency risks, and AI-generation markers.
    Passing scans receive a cryptographically signed compliance certificate.
  contact:
    name: SENTINEL Team
    email: sentinel@archagents.dev
  license:
    name: Proprietary
    identifier: LicenseRef-SENTINEL-1.0
servers:
  - url: https://api.sentinel.example.com
    description: Production
  - url: https://api-staging.sentinel.example.com
    description: Staging
paths:
  /v1/scans:
    get:
      operationId: listScans
      summary: List scans
      description: |
        List scans with filtering by status and project, with pagination support.
      tags: [Scans]
      security:
        - hmacAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, scanning, completed, failed]
          description: Filter by scan status
        - name: projectId
          in: query
          schema:
            type: string
          description: Filter by project ID
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated list of scans
          content:
            application/json:
              schema:
                type: object
                required: [scans, total, limit, offset]
                properties:
                  scans:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScanRecord"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: submitScan
      summary: Submit a scan
      description: |
        Submit a code diff for compliance scanning. Returns a scan ID and
        poll URL to check status. The scan runs asynchronously through the
        agent pipeline (security, license, quality, policy, dependency, AI-detection).
      tags: [Scans]
      security:
        - hmacAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanSubmission"
      responses:
        "202":
          description: Scan accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid HMAC signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/scans/{id}:
    get:
      operationId: getScan
      summary: Get a scan
      description: |
        Retrieve a single scan by its ID, including associated certificate
        and finding count.
      tags: [Scans]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Scan ID
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/scans/{id}/poll:
    get:
      operationId: pollScan
      summary: Poll scan status
      description: |
        Poll the status of a previously submitted scan. Returns the current
        status and, when completed, the full assessment summary.
      tags: [Scans]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Scan ID returned from POST /v1/scans
      responses:
        "200":
          description: Current scan status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/findings:
    get:
      operationId: listFindings
      summary: List findings
      description: |
        List findings across scans with filtering and pagination.
      tags: [Findings]
      security:
        - hmacAuth: []
      parameters:
        - name: scanId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by scan ID
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
          description: Filter by severity
        - name: type
          in: query
          schema:
            type: string
            enum: [security, license, quality, policy, dependency, ai-detection]
          description: Filter by finding type
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated list of findings
          content:
            application/json:
              schema:
                type: object
                required: [findings, total, limit, offset]
                properties:
                  findings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Finding"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/findings/{id}:
    get:
      operationId: getFinding
      summary: Get a finding
      description: |
        Retrieve a single finding by its ID.
      tags: [Findings]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Finding ID
      responses:
        "200":
          description: Finding details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Finding not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/certificates:
    get:
      operationId: listCertificates
      summary: List certificates
      description: |
        List compliance certificates with filtering and pagination.
      tags: [Certificates]
      security:
        - hmacAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
          description: Filter by project ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pass, provisional, fail]
          description: Filter by certificate verdict status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated list of certificates
          content:
            application/json:
              schema:
                type: object
                required: [certificates, total, limit, offset]
                properties:
                  certificates:
                    type: array
                    items:
                      $ref: "#/components/schemas/Certificate"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/certificates/{id}:
    get:
      operationId: getCertificate
      summary: Get a certificate
      description: |
        Retrieve a single compliance certificate by its ID.
      tags: [Certificates]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Certificate ID
      responses:
        "200":
          description: Certificate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/certificates/{id}/verify:
    post:
      operationId: verifyCertificate
      summary: Verify certificate signature
      description: |
        Verify the cryptographic signature of a compliance certificate.
        Returns whether the certificate signature is valid.
      tags: [Certificates]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Certificate ID
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                required: [valid, certificateId]
                properties:
                  valid:
                    type: boolean
                    description: Whether the certificate signature is valid
                  certificateId:
                    type: string
                    format: uuid
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/certificates/{id}/revoke:
    post:
      operationId: revokeCertificate
      summary: Revoke a certificate
      description: |
        Revoke a previously issued compliance certificate. This is an
        irreversible action that updates the certificate status and creates
        an audit log entry.
      tags: [Certificates]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Certificate ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  description: Reason for revocation
      responses:
        "200":
          description: Certificate revoked
          content:
            application/json:
              schema:
                type: object
                required: [id, status, revokedAt]
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [revoked]
                  revokedAt:
                    type: string
                    format: date-time
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/projects:
    get:
      operationId: listProjects
      summary: List projects
      description: |
        List all projects in the organization, including scan counts.
      tags: [Projects]
      security:
        - hmacAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/projects/{id}:
    get:
      operationId: getProject
      summary: Get a project
      description: |
        Retrieve a single project by ID, including scan count and the
        10 most recent scans with their certificates.
      tags: [Projects]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Project ID
      responses:
        "200":
          description: Project details with recent scans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/projects/{id}/findings:
    get:
      operationId: listProjectFindings
      summary: List project findings
      description: |
        List findings for a specific project with filtering and pagination.
        Findings are linked to the project through their parent scan.
      tags: [Findings]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Project ID
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated list of project findings
          content:
            application/json:
              schema:
                type: object
                required: [findings, total, limit, offset]
                properties:
                  findings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Finding"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/policies/{id}:
    put:
      operationId: updatePolicy
      summary: Update a policy
      description: Update an existing compliance policy.
      tags: [Policies]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "200":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePolicy
      summary: Delete a policy
      description: Delete an existing compliance policy.
      tags: [Policies]
      security:
        - hmacAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Policy deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ScanRecord:
      type: object
      required: [id, projectId, commitHash, branch, author, status, startedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
        commitHash:
          type: string
        branch:
          type: string
        author:
          type: string
        status:
          type: string
          enum: [pending, scanning, completed, failed]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        certificate:
          $ref: "#/components/schemas/Certificate"
        _count:
          type: object
          properties:
            findings:
              type: integer
...
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            scans:
              type: array
              items:
                $ref: "#/components/schemas/ScanRecord"
              description: The 10 most recent scans for this project
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          description: Additional error context
  securitySchemes:
    hmacAuth:
      type: apiKey
      in: header
      name: X-Sentinel-Signature
      description: |
        HMAC-SHA256 signature of the request body using the shared secret.
        Format: sha256=<hex-encoded-signature>